+++ src/components/StatusBarIOS.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { StatusBar } from 'react-native'
+import { Container } from './styles/StatusBarIOS.styles'
+
+export default function StatusBarIOS({ backgroundColor, ...props }: { backgroundColor: string }) {
+
+    return (
+        <Container style={[{ backgroundColor }]}>
+            <StatusBar translucent backgroundColor={backgroundColor} {...props} />
+        </Container>
+    )
+}