+++ src/navigation/index.tsx
@@ -3,10 +3,9 @@ import { DarkTheme, DefaultTheme, NavigationContainer, useNavigation } from '@re
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import * as React from 'react'
 import { useEffect, useState } from 'react'
-import { ColorSchemeName, Dimensions, Image, useWindowDimensions } from 'react-native'
-import { IconButton } from 'react-native-paper'
+import { ColorSchemeName, Dimensions, Image, Platform, useWindowDimensions } from 'react-native'
 import { RootStackParamList, RootStackScreenProps, RootTabParamList, RootTabScreenProps } from '../../types'
-import { ModalImage } from '../components/styles/OverviewHeader.styles'
+import SnackbarMessage from '../components/Snackbar'
 import TabBarIcon from '../components/TabBarIcon'
 import Colors from '../constants/Colors'
 import useColorScheme from '../hooks/useColorScheme'
@@ -22,9 +21,8 @@ import ScanTransactionScreen from '../screens/ScanTransactionScreen'
 import SignInScreen from '../screens/SignInScreen'
 import WalletTabScreen from '../screens/WalletTabScreen'
 import LocalStorageService from '../services/LocalStorage.service'
-import { backgroundImage, logoTitle } from '../utils/images'
+import { backgroundImage } from '../utils/images'
 import LinkingConfiguration from './LinkingConfiguration'
-import SnackbarMessage from '../components/Snackbar'
 
 const navTheme = {
   ...DefaultTheme,
@@ -93,31 +91,23 @@ function RootNavigator(): any {
           <Stack.Screen name="OnBoarding" component={OnboardingScreen} options={{ headerShown: false }} />
         )}
         <Stack.Screen name="SignIn" component={SignInScreen} options={{ headerShown: false }} />
-        <Stack.Screen name="ScanKey" component={ScanKeyScreen} options={{ headerTitle: 'Back to Sign In', headerTransparent: true }} />
-        <Stack.Screen name="ImportAccount" component={ImportAccountScreen} options={{ headerTitle: 'Back to Sign In' }} />
+        <Stack.Screen name="ScanKey" component={ScanKeyScreen} options={{ headerBackTitle: 'Back to Sign In', headerTransparent: true, headerTitle: '' }} />
+        <Stack.Screen name="ImportAccount" component={ImportAccountScreen} options={{ headerTitle: '' }} />
         <Stack.Screen name="Root" component={BottomTabNavigator} options={{ headerShown: false }} />
         <Stack.Screen name="NotFound" component={NotFoundScreen} options={{ title: 'Oops!' }} />
-        <Stack.Screen name="ScanTransaction" component={ScanTransactionScreen} options={{ headerTitle: 'Go back', }} />
-        <Stack.Group screenOptions={{ presentation: 'modal' }}>
+        <Stack.Screen name="ScanTransaction" component={ScanTransactionScreen} options={{ headerBackTitle: 'Go back', headerTransparent: true, headerTitle: '' }} />
+        <Stack.Group>
           <Stack.Screen name="Modal" component={ModalScreen}
             options={({ navigation }: RootStackScreenProps<'Modal'>) => ({
-              headerBackVisible: false,
-              headerLeft: () => (<ModalImage testID="logo-title" source={logoTitle} />),
-              headerTitleStyle: { color: 'transparent' },
-              headerStyle: { backgroundColor: '#ffffff' },
-              headerRight: () => (
-                <IconButton
-                  icon='close'
-                  color={Colors[colorScheme].tint}
-                  size={25}
-                  onPress={() => navigation.navigate('Root')} />
-              )
+              headerShown: false,
             })} />
           <Stack.Screen name="Profile" component={ProfileScreen}
             options={{
-              headerTitle: 'Profile',
+              headerTitle: '',
+              headerBackTitle: 'Profile',
               headerStyle: { backgroundColor: '#ffffff' }
             }} />
+
         </Stack.Group>
       </Stack.Navigator>
     )
@@ -137,7 +127,7 @@ function BottomTabNavigator() {
       initialLayout={{ width: Dimensions.get('window').width, height: Dimensions.get('window').height }}
       screenOptions={{
         tabBarActiveTintColor: Colors[colorScheme].tint,
-        tabBarStyle: { height: 65 },
+        tabBarStyle: { height: 75 },
       }}
     >
       <Tab.Screen