+++ src/screens/WalletTabScreen.tsx
@@ -1,9 +1,9 @@
+import { StatusBar } from 'expo-status-bar'
 import React, { useEffect, useState } from 'react'
-import { StatusBar } from 'react-native'
 import { Card, Paragraph, Title } from 'react-native-paper'
 import { RootTabScreenProps } from '../../types'
 import OverviewHeader from '../components/OverviewHeader'
-import QRButton from '../components/QRIcon'
+import QRButton from '../components/QRButton'
 import Spinner from '../components/Spinner'
 import { StyledImage } from '../components/styles/OverviewHeader.styles'
 import { AmountContainer, BottomCard, BottomCardsContainer, OverviewContainer, TopCard, TopCardsContainer } from '../components/styles/WalletTabScreen.styles'
@@ -12,6 +12,8 @@ import CoinMarketCapService from '../services/CoinMarketCap.service'
 import LocalStorageService from '../services/LocalStorage.service'
 import { formatNumber } from '../utils/formatNumber'
 import { logoTitle } from "../utils/images"
+import { FieldName, BlueText, RedText, Amount } from '../components/styles/WalletTabScreen.styles'
+import StatusBarIOS from '../components/StatusBarIOS'
 
 export default function WalletTabScreen({ navigation, route }: RootTabScreenProps<'Wallet'>) {
 
@@ -76,31 +78,30 @@ export default function WalletTabScreen({ navigation, route }: RootTabScreenProp
                 <Spinner />
                 :
                 <>
-                    <OverviewHeader
-                        icon={"menu"}
-                        onPress={() => navigation.navigate('Modal')}
-                        input={<StyledImage testID="logo-title" source={logoTitle} />} />
+                    <StatusBarIOS backgroundColor={'#ffffff'} />
                     <OverviewContainer>
-
-                        <StatusBar backgroundColor={'#ffffff'} />
+                        <OverviewHeader
+                            icon={"menu"}
+                            onPress={() => navigation.navigate('Modal')}
+                            input={<StyledImage testID="logo-title" source={logoTitle} />} />
 
                         <TopCardsContainer>
 
                             <TopCard>
                                 <Card.Content style={{ borderRadius: 80 }}>
-                                    <Paragraph>Regular account</Paragraph>
+                                    <FieldName>Regular account</FieldName>
                                     <AmountContainer>
-                                        <Title>{formatNumber(regular)}</Title><Paragraph>LTO</Paragraph>
+                                        <Amount>{formatNumber(regular)}</Amount><Paragraph>LTO</Paragraph>
                                     </AmountContainer>
-                                    <Paragraph>Equivalent to {formatNumber(change)}$</Paragraph>
+                                    <BlueText>Equivalent to {formatNumber(change)}$</BlueText>
                                 </Card.Content>
                             </TopCard>
 
                             <TopCard>
                                 <Card.Content>
-                                    <Paragraph>Prize</Paragraph>
-                                    <Title>{price?.toFixed(3)}</Title>
-                                    <Paragraph>{percent_change_24h?.toFixed(2)}%(last 24h)</Paragraph>
+                                    <FieldName>Prize</FieldName>
+                                    <Amount>{price?.toFixed(3)}$</Amount>
+                                    <RedText>{percent_change_24h?.toFixed(2)}%(last 24h)</RedText>
                                 </Card.Content>
                             </TopCard>
 
@@ -110,18 +111,18 @@ export default function WalletTabScreen({ navigation, route }: RootTabScreenProp
 
                             <BottomCard>
                                 <Card.Content>
-                                    <Paragraph>Generating</Paragraph>
+                                    <FieldName>Generating</FieldName>
                                     <AmountContainer>
-                                        <Title>{formatNumber(generating)}</Title><Paragraph>LTO</Paragraph>
+                                        <Amount>{formatNumber(generating)}</Amount><Paragraph>LTO</Paragraph>
                                     </AmountContainer>
                                 </Card.Content>
                             </BottomCard>
 
                             <BottomCard >
                                 <Card.Content>
-                                    <Paragraph>Available</Paragraph>
+                                    <FieldName>Available</FieldName>
                                     <AmountContainer>
-                                        <Title>{formatNumber(available)}</Title><Paragraph>LTO</Paragraph>
+                                        <Amount>{formatNumber(available)}</Amount><Paragraph>LTO</Paragraph>
                                     </AmountContainer>
                                 </Card.Content>
                             </BottomCard>
@@ -132,25 +133,24 @@ export default function WalletTabScreen({ navigation, route }: RootTabScreenProp
 
                             <BottomCard >
                                 <Card.Content>
-                                    <Paragraph>Effective</Paragraph>
+                                    <FieldName>Effective</FieldName>
                                     <AmountContainer>
-                                        <Title>{formatNumber(change)}</Title><Paragraph>$</Paragraph>
+                                        <Amount>{formatNumber(change)}</Amount><Paragraph>$</Paragraph>
                                     </AmountContainer>
                                 </Card.Content>
                             </BottomCard>
 
                             <BottomCard >
                                 <Card.Content>
-                                    <Paragraph>Unbonding</Paragraph>
-                                    <Title>{formatNumber(unbonding)}</Title>
+                                    <FieldName>Unbonding</FieldName>
+                                    <Amount>{formatNumber(unbonding)}</Amount>
                                 </Card.Content>
                             </BottomCard>
 
                         </BottomCardsContainer>
 
-
+                        <QRButton onPress={() => navigation.navigate('ScanTransaction')} />
                     </OverviewContainer>
-                    <QRButton onPress={() => navigation.navigate('ScanTransaction')} />
                 </>
             }
         </>